Change function declaration syntax
The syntax to declare a function is now:

fun(x, y) -> x + y

instead of:

fun x y -> x + y

The change was made as it reflects the change to make the language more
imperative, rather than functional.

The tests have been updated with this new syntax, and the old examples
have been removed.