Add new syntax for if-else blocks
When an if-else block is declared, it now must end with the "end"
keyword. This is to minimise shift/reduce conflicts in Yip, as was some
ambiguity as to what part of a statement was in the function. It's the
same syntax as Lua uses to end functions.

Also added the files I forgot to include in the previous commit.