Add blocks to the language
Added a block construct to the language, allowing you to define multiple
statements in a single expression. Blocks are declared by enclosing
statements between curly braces.

For instance:
{
	x = 3
	x + 17
}

Blocks return the value of the last statement within them.

If statements and functions now only enclose a single expression, so
blocks allow one to use multiple statements in a row.