Add external functions & repeat keyword
External functions can now be called from the language. They must first
be defined using the "extern" keyword. You can then call them in normal
code.

The repeat keyword keyword has also been added to the language. It can
only be placed inside a block, and when executed it jumps back to the
start of the block. It allows for loops to be implemented in the
language.