If statements don't use phi nodes
The code generation for if statements has been changed in order for the
repeat keyword to work properly.

Performing a repeat in an if statement would result in invalid code.
There would be extra cases that the phi node generated for the if
statement would not handle.

Now if statements output in a mutable stack variable, so there is no
need to use the phi statement anymore.